feat: Add `std/date` and `std/timer` modules, introduce `Int64` type, enhance string concatenation, and update function value representation.

+++ src/compiler/ast/mod.rs
@@ -35,7 +35,7 @@ impl TypeExpr {
 
 #[derive(Debug, Clone)]
 pub enum Expr {
-    Number(i32, Span),
+    Number(i64, Span),
     StringLiteral(String, Span),
     Variable(String, Span),
     BinaryOp(Box<Expr>, String, Box<Expr>, Span),
